requirements.txt
pydicom>=1.4.2
dicompyler-core[image]>=0.5.2
wxPython>=4.0.0b2
matplotlib>=1.3,<2.2
numpy>=1.13.1
Pypubsub>=4.0.3
wxmplot>=0.9.40
https://github.com/darcymason/pydicom/archive/master.zip
